Don't suspend the thread if the DAC is already setup.
This fixes a fairly big issue with a race condition. Normally, the main
thread will sleep until the DAC is spun up, and then calculate the
drawables. On occasion, the DAC would get spun up first and the main
thread would be waiting forever, foreverever, foreverever. Also notify
the mutex if the DAC fails just to not lock things up