Have fingerable elements handle touch events
When a finger is first pressed, we propagate the touch event down the
views until something matches. When (If) something matches, we return it
and set it to fingersById. This way, next time the finger id is seen, we
deliver the touch event directly to the fingerable. This gets us the
benefit of a finger is pressed on an element, and as the finger moves
away from that element, the element still responds to events (e.g.
knobs).