Use angle deltas to determine UI for knobs
This seems to make a lot of sense. If you increase the angle, the knob
should increase along side. This allows for people to use them like
normal knobs. If the angle stays the same, then we default to the
distance, which allows people to handle them like sliders. There are
some inconsistencies with the arctan function in this implementation,
but overall it's Solid Enough For Now.