Create a basic templating engine for Drawables
Essentially, all drawables have a property that determines whether they
should clear floating and create a new row. As Drawables are added, they
are added to the current row (up until the last seen Drawable that
clears). Columns are equally spaced within a row. This is just a very
basic table-based templating engine, but should be Just Enough for now.