Encapsulate the finger/oscillator combo into one Fingerable/Drawable
ComplexOscs handle sound primarily, so they shouldn't really have any
concern with drawing themselves. Consequently, we make a
Fingerable/Drawable class that handles the display of the fingers and
manipulating the oscillator on touch events. This fits our scheme a
little better since everything touchable is now becoming a Fingerable
element. We have to do some unfortunate hacks, like pass the SauceView
around so that we can scale x/y values, but on the whole, this is much
nicer and fits our current pattern.