Modify Knob UI to be a little simpler
Rather than go from the center, we go from the last point the user was
at. We also remove the "same angle" case because it seems to just cause
issues. End result of all this is that Knob UIs are really sucky to
mimic on a 2d surface. Starting to prefer Sliders for that functionality