Pause the dac ticking when the app is paused and we're not looping
paused = the app goes into the background. when the thread is ticking in
the background, it gets put on low priority which results in stuttering
and pops in the speaker. So if we're not looping, then chances are we
don't care about keeping the audio playing in the background, so we can
pause the dac thread.