properly fill the silent buffer with short based "0" value
and seed that buffer as the first thing on the audio track. trying to get rid of pops, and it seems like one main reason would be that the audio track bounces from 0 to a huge value. "0" is our minimum value since we use shorts. our baseline value is half of max_short, so this should be accurate. seeding it to start just seems like a good idea.